New zu.json interface
- Reviewed the json module interface to address stringified json
schema as zu.json.stringified()
- Exported inferred types under subpath
