+++ src/json/index.ts
@@ -1,0 +1,13 @@
+import { json, literal, stringifiedJson } from "./json.js";
+
+const module = (() => {
+  function module() {
+    return json();
+  }
+  module.stringified = stringifiedJson;
+  module.literal = literal;
+
+  return module;
+})();
+
+export { module as json };
