+++ src/json/types.ts
@@ -1,0 +1,1 @@
+export type { Literal, Json } from "./json.js";
